refactor(robots): move robots config to utilities, single source for /robots.txt
Next 15 disallows both app/robots.ts (metadata file) and
app/robots.txt/route.ts (route handler) — the build errors out with
"Conflicting route and metadata at /robots.txt". The route handler
wins because it carries the NLWeb Schemamap directive, so move the
typed rules out of app/.

- Move app/robots.ts -> utilities/robots-config.ts (default export
  preserved)
- Route handler in app/robots.txt/route.ts imports from there
- Test in __tests__/app/robots.test.ts re-points the same import
